+++ src/examples/telemetry_buffer_manager_example.cpp
@@ -7,7 +7,7 @@
  */
 
 
-#include <yarp/telemetry/experimental/BufferManager.h>
+#include <robometry/BufferManager.h>
 
 #include <iostream>
 #include <iomanip>
@@ -22,16 +22,16 @@ constexpr size_t n_samples{3};
 int main()
 {
     // The inputs to the API are defined in the BufferConfig structure
-    yarp::telemetry::experimental::BufferConfig bufferConfig;
+    robometry::BufferConfig bufferConfig;
 
     // We use the default config, setting only the number of samples (no auto/periodic saving)
     bufferConfig.n_samples = n_samples;
 
-    yarp::telemetry::experimental::BufferManager<int32_t> bm(bufferConfig);
+    robometry::BufferManager<int32_t> bm(bufferConfig);
     bm.setFileName("buffer_manager_test");
     auto ok{false};
-    yarp::telemetry::experimental::ChannelInfo var_one{ "one", {1,1} };
-    yarp::telemetry::experimental::ChannelInfo var_two{ "two", {1,1} };
+    robometry::ChannelInfo var_one{ "one", {1,1} };
+    robometry::ChannelInfo var_two{ "two", {1,1} };
 
     ok = bm.addChannel(var_one);
     ok = ok && bm.addChannel(var_two);
@@ -54,9 +54,9 @@ int main()
     // now we test our API with the auto_save option enabled.
     bufferConfig.auto_save = true;
 
-    yarp::telemetry::experimental::BufferManager<int32_t> bm_m(bufferConfig);
+    robometry::BufferManager<int32_t> bm_m(bufferConfig);
     bm_m.setFileName("buffer_manager_test_matrix");
-    std::vector<yarp::telemetry::experimental::ChannelInfo> vars{ { "one",{2,3} },
+    std::vector<robometry::ChannelInfo> vars{ { "one",{2,3} },
                                                     { "two",{3,2} } };
 
     ok = bm_m.addChannels(vars);
@@ -75,7 +75,7 @@ int main()
     bufferConfig.channels = { {"one",{4,1}}, {"two",{4,1}} };
     bufferConfig.filename = "buffer_manager_test_vector";
 
-    yarp::telemetry::experimental::BufferManager<double> bm_v(bufferConfig);
+    robometry::BufferManager<double> bm_v(bufferConfig);
 
     for (int i = 0; i < 10; i++) {
         bm_v.push_back({ i+1.0, i+2.0, i+3.0, i+4.0  }, "one");
@@ -87,7 +87,7 @@ int main()
     bufferConfig.channels = { {"struct1::one",{4,1}}, {"struct1::two",{4,1}}, {"struct2::one",{4,1}} };
     bufferConfig.filename = "buffer_manager_test_nested_vector";
 
-    yarp::telemetry::experimental::BufferManager<double> bm_ns(bufferConfig);
+    robometry::BufferManager<double> bm_ns(bufferConfig);
 
     for (int i = 0; i < 10; i++) {
         bm_ns.push_back({ i+1.0, i+2.0, i+3.0, i+4.0  }, "struct1::one");