+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Renamed the repo robometry instead of yarp-telemetry.
+- Renamed `YARP_telemetry` in `robometry`.
+
 ## [0.5.1] - 2022-05-06
 
 - Added the method `setSaveCallback` for adding an additiona function invoked with